test: split and refactor charm tests
