+++ tests/unit/test_charm.py
@@ -17,109 +17,17 @@ import requests
 import yaml
 
 import jenkins
-import precondition
 import state
 import timerange
 from charm import JenkinsK8sOperatorCharm, ReconcileBlockedError
 
-from .helpers import ACTIVE_STATUS_NAME, BLOCKED_STATUS_NAME, WAITING_STATUS_NAME
-from .types_ import Harness, HarnessWithContainer
+from .helpers import ACTIVE_STATUS_NAME
+from .types_ import HarnessWithContainer
 
-
-@pytest.mark.parametrize(
-    "charm_config",
-    [pytest.param({"restart-time-range": "-2"}, id="invalid restart-time-range")],
-)
-def test__on_config_changed_invalid_config_sets_blocked(
-    harness_container: HarnessWithContainer, charm_config: dict[str, str]
-):
-    """
-    arrange: given an invalid charm configuration.
-    act: when the config-changed handler is triggered.
-    assert: the charm falls into blocked status.
-    """
-    harness_container.harness.update_config(charm_config)
-    harness_container.harness.begin()
-
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness_container.harness.charm)
-    jenkins_charm._reconcile(MagicMock(spec=ops.ConfigChangedEvent))
-
-    assert jenkins_charm.unit.status.name == BLOCKED_STATUS_NAME, "unit should be in BlockedStatus"
-
-
-@pytest.mark.parametrize(
-    "event_spec",
-    [
-        pytest.param(ops.StorageAttachedEvent, id="_on_jenkins_home_storage_attached"),
-        pytest.param(ops.PebbleReadyEvent, id="_on_jenkins_pebble_ready"),
-        pytest.param(ops.UpdateStatusEvent, id="_on_update_status"),
-    ],
-)
-def test_workload_not_ready(harness: Harness, event_spec: type):
-    """
-    arrange: given a charm with no container ready.
-    act: when an event hook is fired.
-    assert: the charm falls into waiting status.
-    """
-    harness.add_storage(state.JENKINS_HOME_STORAGE_NAME, count=1, attach=True)
-    harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-    mock_event = MagicMock(spec=event_spec)
-
-    if event_spec == ops.PebbleReadyEvent:
-        jenkins_charm._on_jenkins_pebble_ready(mock_event)
-    else:
-        jenkins_charm._reconcile(mock_event)
-
-    assert jenkins_charm.unit.status.name == WAITING_STATUS_NAME
-
-
-@pytest.mark.parametrize(
-    "event_spec",
-    [
-        pytest.param(ops.StorageAttachedEvent, id="_on_jenkins_home_storage_attached"),
-        pytest.param(ops.PebbleReadyEvent, id="_on_jenkins_pebble_ready"),
-        pytest.param(ops.UpdateStatusEvent, id="_on_update_status"),
-    ],
-)
-def test_storage_not_ready(harness: Harness, event_spec: type):
-    """
-    arrange: given a charm with no storage ready.
-    act: when an event hook is fired.
-    assert: the charm falls into waiting status.
-    """
-    harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-    container = harness.model.unit.get_container("jenkins")
-    harness.set_can_connect(container, True)
-    mock_event = MagicMock(spec=event_spec)
-
-    if event_spec == ops.PebbleReadyEvent:
-        jenkins_charm._on_jenkins_pebble_ready(mock_event)
-    else:
-        jenkins_charm._reconcile(mock_event)
-
-    assert jenkins_charm.unit.status.name == WAITING_STATUS_NAME
-
-
-def test__get_state_returns_none_on_invalid_config(harness: Harness):
-    """
-    arrange: given an invalid charm config.
-    act: when _get_state() is called.
-    assert: unit is BlockedStatus and None is returned.
-    """
-    harness.update_config({"restart-time-range": "-2"})
-    harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-
-    with patch("state.State.from_charm") as from_charm_mock:
-        from_charm_mock.side_effect = state.CharmConfigInvalidError("bad config")
-        result = jenkins_charm._get_state()
-
-    assert result is None
-    assert jenkins_charm.unit.status.name == BLOCKED_STATUS_NAME
-    assert jenkins_charm.unit.status.message == "bad config"
-
+# Reconcile/config and event-routing tests moved to:
+# - tests/unit/test_charm_reconcile.py
+# - tests/unit/test_charm_events.py
+# for module simplicity
 
 def test__on_jenkins_pebble_ready_get_version_error(
     harness_container: HarnessWithContainer,
@@ -297,142 +205,7 @@ def test__on_update_status(
         jenkins_charm._reconcile_plugins(harness_container.container, dummy_state)
 
 
-def test_calculate_env(harness: Harness):
-    """
-    arrange: given a charm.
-    act: when calculate_env is called.
-    assert: expected environment variable mapping dictionary is returned.
-    """
-    harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-    with patch.object(jenkins_charm, "_get_ingress_path", return_value=""):
-        env = jenkins_charm.calculate_env()
-
-    assert env == {
-        "JENKINS_HOME": str(jenkins.JENKINS_HOME_PATH),
-        "JENKINS_PREFIX": "",
-        "CASC_JENKINS_CONFIG": str(jenkins.JCASC_CONFIG_PATH),
-        "JENKINS_ADMIN_PASSWORD": "",
-    }
-
-
-def test__on_config_changed_invalid_config_blocked(
-    harness_container: HarnessWithContainer,
-):
-    """
-    arrange: given State.from_charm raises CharmConfigInvalidError during config-changed.
-    act: when _on_config_changed is invoked.
-    assert: unit falls into BlockedStatus and no container actions are performed.
-    """
-    # Start charm normally so __init__ completes and observers are registered.
-    harness = harness_container.harness
-    harness.begin()
-
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-
-    with (
-        patch("state.State.from_charm") as from_charm_mock,
-        patch.object(harness_container.container, "add_layer") as add_layer_mock,
-        patch.object(harness_container.container, "replan") as replan_mock,
-        patch.object(harness_container.container, "restart") as restart_mock,
-    ):
-        from_charm_mock.side_effect = state.CharmConfigInvalidError("bad sysprops")
-
-        jenkins_charm._reconcile(MagicMock(spec=ops.ConfigChangedEvent))
-
-        assert jenkins_charm.unit.status.name == BLOCKED_STATUS_NAME
-        assert jenkins_charm.unit.status.message == "bad sysprops"
-        add_layer_mock.assert_not_called()
-        replan_mock.assert_not_called()
-        restart_mock.assert_not_called()
-
-
-def test__on_config_changed_relation_data_invalid_raises(
-    harness_container: HarnessWithContainer,
-):
-    """
-    arrange: given State.from_charm raises CharmRelationDataInvalidError.
-    act: when _on_config_changed is invoked.
-    assert: a RuntimeError is raised by the handler.
-    """
-    harness = harness_container.harness
-    harness.begin()
-
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-
-    with patch("state.State.from_charm") as from_charm_mock:
-        from_charm_mock.side_effect = state.CharmRelationDataInvalidError("bad relation")
-
-        with pytest.raises(RuntimeError):
-            jenkins_charm._reconcile(MagicMock(spec=ops.ConfigChangedEvent))
-
-
-def test__on_config_changed_precondition_waits_and_defers(
-    harness_container: HarnessWithContainer,
-):
-    """
-    arrange: given precondition.check indicates not ready.
-    act: when _on_config_changed is invoked.
-    assert: unit is in WaitingStatus and the event is deferred.
-    """
-    harness = harness_container.harness
-    harness.begin()
-
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-    event = MagicMock(spec=ops.ConfigChangedEvent)
-
-    with (
-        patch("state.State.from_charm", wraps=state.State.from_charm),
-        patch("precondition.check") as check_mock,
-        patch.object(harness_container.container, "add_layer") as add_layer_mock,
-        patch.object(harness_container.container, "replan") as replan_mock,
-        patch.object(harness_container.container, "restart") as restart_mock,
-    ):
-        check_mock.return_value = precondition._CheckResult(success=False, reason="not ready")
-
-        jenkins_charm._reconcile(event)
-
-        assert jenkins_charm.unit.status.name == WAITING_STATUS_NAME
-        assert jenkins_charm.unit.status.message == "not ready"
-        event.defer.assert_called_once_with()
-        add_layer_mock.assert_not_called()
-        replan_mock.assert_not_called()
-        restart_mock.assert_not_called()
-
-
-def test__on_config_changed_success_replans_and_restarts(
-    harness_container: HarnessWithContainer,
-):
-    """
-    arrange: given precondition.check success and a valid pebble layer builder.
-    act: when _on_config_changed is invoked.
-    assert: container.add_layer and container.replan are called.
-    """
-    harness = harness_container.harness
-    harness.begin()
-
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness.charm)
-
-    with (
-        patch("state.State.from_charm", wraps=state.State.from_charm),
-        patch("precondition.check") as check_mock,
-        patch("pebble.get_pebble_layer") as layer_mock,
-        patch.object(harness_container.container, "add_layer") as add_layer_mock,
-        patch.object(harness_container.container, "replan") as replan_mock,
-        patch.object(harness_container.container, "get_plan") as get_plan_mock,
-        patch.object(jenkins_charm, "_reconcile_jcasc"),
-        patch.object(jenkins_charm, "_reconcile_plugins"),
-    ):
-        check_mock.return_value = precondition._CheckResult(success=True, reason=None)
-        # Minimal viable layer for add_layer
-        layer_mock.return_value = ops.pebble.Layer({"services": {"jenkins": {}}})
-        # Return empty plan so services differ and replan is triggered
-        get_plan_mock.return_value = ops.pebble.Plan("")
-
-        jenkins_charm._reconcile(MagicMock(spec=ops.ConfigChangedEvent))
-
-        add_layer_mock.assert_called_once()
-        replan_mock.assert_called_once()
+## Reconcile/config tests moved to tests/unit/test_charm_reconcile.py
 
 
 def test__remove_unlisted_plugins_requires_state(
@@ -452,103 +225,7 @@ def test__remove_unlisted_plugins_requires_state(
         reconcile_plugins(harness_container.container)
 
 
-@pytest.mark.parametrize(
-    "event_type",
-    [
-        pytest.param(ops.RelationJoinedEvent, id="joined"),
-        pytest.param(ops.RelationDepartedEvent, id="departed"),
-        pytest.param(ops.RelationChangedEvent, id="changed"),
-    ],
-)
-def test__agent_relation_handlers_reconcile_agents(
-    harness_container: HarnessWithContainer, event_type: type[ops.EventBase]
-):
-    """
-    arrange: given a started charm and a valid derived state.
-    act: when an agent relation event triggers _reconcile.
-    assert: reconcile executes without error (delegates internally to _reconcile_agents).
-    """
-    harness_container.harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness_container.harness.charm)
-    event = MagicMock(spec=event_type)
-
-    with patch.object(jenkins_charm, "_reconcile_agents") as reconcile_agents_mock:
-        jenkins_charm._reconcile(event)
-
-    reconcile_agents_mock.assert_called_once()
-
-
-@pytest.mark.parametrize(
-    "event_spec",
-    [
-        pytest.param(ops.EventBase, id="ready"),
-        pytest.param(ops.EventBase, id="revoked"),
-    ],
-)
-def test__agent_discovery_ingress_handlers_reconfigure_agents(
-    harness_container: HarnessWithContainer, event_spec: type
-):
-    """
-    arrange: given a started charm and a valid derived state.
-    act: when an agent discovery ingress event triggers _reconcile.
-    assert: reconcile executes without error (delegates internally to _reconcile_agent_discovery).
-    """
-    harness_container.harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness_container.harness.charm)
-    event = MagicMock(spec=event_spec)
-
-    with patch.object(jenkins_charm, "_reconcile_agent_discovery") as reconcile_discovery_mock:
-        jenkins_charm._reconcile(event)
-
-    reconcile_discovery_mock.assert_called_once()
-
-
-def test__upgrade_charm_reconciles_storage_and_agents(
-    harness_container: HarnessWithContainer,
-):
-    """
-    arrange: given a started charm and a valid derived state.
-    act: when the upgrade-charm event triggers _reconcile.
-    assert: _reconcile_storage is called (UpgradeCharmEvent triggers storage reconciliation).
-    """
-    harness_container.harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness_container.harness.charm)
-    event = MagicMock(spec=ops.UpgradeCharmEvent)
-
-    with patch.object(jenkins_charm, "_reconcile_storage") as reconcile_storage_mock:
-        jenkins_charm._reconcile(event)
-
-    reconcile_storage_mock.assert_called_once()
-
-
-@pytest.mark.parametrize(
-    "event_type",
-    [
-        pytest.param(
-            ops.RelationJoinedEvent,
-            id="joined",
-        ),
-        pytest.param(
-            ops.RelationDepartedEvent,
-            id="departed",
-        ),
-    ],
-)
-def test__auth_proxy_relation_handlers_delegate(
-    harness_container: HarnessWithContainer,
-    event_type: type[ops.EventBase],
-):
-    """
-    arrange: given a started charm and a valid derived state.
-    act: when an auth-proxy relation event triggers _reconcile.
-    assert: reconcile executes without error (delegates internally to _reconcile_auth_proxy).
-    """
-    harness_container.harness.begin()
-    jenkins_charm = typing.cast(JenkinsK8sOperatorCharm, harness_container.harness.charm)
-    event = MagicMock(spec=event_type)
-
-    with patch.object(jenkins_charm, "_reconcile_auth_proxy"):
-        jenkins_charm._reconcile(event)
+## Event-routing tests moved to tests/unit/test_charm_events.py
 
 
 VALID_JCASC_CONFIG = {
